requirements.txt
pandas>=1.5.0
streamlit>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0
